[onert] Unify "cpu_common" of cpu and controlflow in "core"
This commit unifies "cpu_common" of cpu and controlflow in "core".
  - Move "cpu_common" from backend dir to "core/backend"
  - Unify controlflow's MemoryManager/Allocator with them of "cpu_common"
