+++ runtime/onert/backend/CMakeLists.txt
@@ -1,8 +1,6 @@
 set(LIB_ONERT_BACKEND_ACL_COMMON onert_backend_acl_common)
-set(LIB_ONERT_BACKEND_CPU_COMMON onert_backend_cpu_common)
 
 add_subdirectory(cpu)
-add_subdirectory(cpu_common)
 add_subdirectory(acl_cl)
 add_subdirectory(acl_neon)
 add_subdirectory(acl_common)
