+++ runtime/onert/backend/cpu/StaticTensorManager.h
@@ -17,7 +17,7 @@
 #ifndef __ONERT_BACKEND_CPU_STATICTENSOR_MANAGER_H__
 #define __ONERT_BACKEND_CPU_STATICTENSOR_MANAGER_H__
 
-#include "MemoryManager.h"
+#include <backend/cpu_common/MemoryManager.h>
 #include "TensorRegistry.h"
 #include "Tensor.h"
