+++ runtime/onert/core/src/backend/cpu_common/Allocator.cc
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-#include "Allocator.h"
+#include "backend/cpu_common/Allocator.h"
 
 #include "util/logging.h"
