+++ runtime/onert/core/src/backend/cpu_common/MemoryManager.cc
@@ -18,7 +18,7 @@
 
 #include <cassert>
 
-#include <MemoryPlannerFactory.h>
+#include "MemoryPlannerFactory.h"
 #include "util/ConfigSource.h"
 
 namespace onert
