+++ runtime/onert/core/src/backend/cpu_common/MemoryPlannerFactory.cc
@@ -14,12 +14,9 @@
  * limitations under the License.
  */
 
-#ifndef __ONERT_BACKEND_CPU_COMMON_MEMORY_PLANNER_FACTORY_H__
-#define __ONERT_BACKEND_CPU_COMMON_MEMORY_PLANNER_FACTORY_H__
+#include "MemoryPlannerFactory.h"
 
 #include "MemoryPlanner.h"
-
-#include <string>
 
 namespace onert
 {
@@ -28,20 +25,29 @@ namespace backend
 namespace cpu_common
 {
 
-class MemoryPlannerFactory
+MemoryPlannerFactory &MemoryPlannerFactory::get()
 {
-public:
-  static MemoryPlannerFactory &get();
+  static MemoryPlannerFactory instance;
+  return instance;
+}
 
-private:
-  MemoryPlannerFactory() = default;
-
-public:
-  IMemoryPlanner *create(const std::string &key);
-};
+IMemoryPlanner *MemoryPlannerFactory::create(const std::string &key)
+{
+  if (key == "FirstFit")
+  {
+    return new FirstFitPlanner;
+  }
+  else if (key == "Bump")
+  {
+    return new BumpPlanner;
+  }
+  else if (key == "WIC")
+  {
+    return new WICPlanner;
+  }
+  return new FirstFitPlanner; // Default Planner
+}
 
 } // namespace cpu_common
 } // namespace backend
 } // namespace onert
-
-#endif // __ONERT_BACKEND_CPU_COMMON_MEMORY_PLANNER_FACTORY_H__
